requirements.txt
keyboard>=0.13.5
pywin32>=306
gspread>=6.0.0
google-auth>=2.28.0
psutil>=5.9.0
comtypes>=1.4.0
PyQt6>=6.6.0
